+++ src/app/mentor/_components/MentorClient/_components/MyMentoSection/index.tsx
@@ -1,0 +1,19 @@
+import { getMentorData } from "@/utils/mockingGetData";
+
+import MentorCard from "@/components/mentor/MentorCard";
+
+const MyMentoSection = () => {
+  const myMentoData = getMentorData();
+
+  return (
+    <>
+      <h2 className="text-lg font-bold text-gray-900">나의 멘토 페이지</h2>
+      <div className="mt-[14px]">
+        {myMentoData.map((mentee) => (
+          <MentorCard key={mentee.id} isMine mentor={mentee} />
+        ))}
+      </div>
+    </>
+  );
+};
+export default MyMentoSection;
